+++ edtrio-client/schema.json
@@ -2659,6 +2659,33 @@
             "deprecationReason": null
           },
           {
+            "name": "deletePoll",
+            "description": "",
+            "args": [
+              {
+                "name": "pollId",
+                "description": "",
+                "type": {
+                  "kind": "NON_NULL",
+                  "name": null,
+                  "ofType": {
+                    "kind": "SCALAR",
+                    "name": "String",
+                    "ofType": null
+                  }
+                },
+                "defaultValue": null
+              }
+            ],
+            "type": {
+              "kind": "OBJECT",
+              "name": "Poll",
+              "ofType": null
+            },
+            "isDeprecated": false,
+            "deprecationReason": null
+          },
+          {
             "name": "updatePoll",
             "description": "",
             "args": [
@@ -2733,7 +2760,34 @@
             "deprecationReason": null
           },
           {
-            "name": "addUserToPollAnswer",
+            "name": "deletePollAnswer",
+            "description": "",
+            "args": [
+              {
+                "name": "pollAnswerId",
+                "description": "",
+                "type": {
+                  "kind": "NON_NULL",
+                  "name": null,
+                  "ofType": {
+                    "kind": "SCALAR",
+                    "name": "String",
+                    "ofType": null
+                  }
+                },
+                "defaultValue": null
+              }
+            ],
+            "type": {
+              "kind": "OBJECT",
+              "name": "PollAnswer",
+              "ofType": null
+            },
+            "isDeprecated": false,
+            "deprecationReason": null
+          },
+          {
+            "name": "addSubmissionToPollAnswer",
             "description": "",
             "args": [
               {
