+++ edtrio-client/src/graphqlOperations/operations.ts
@@ -71,8 +71,8 @@ export const DELETE_POLL_ANSWER = gql`
 `;
 
 export const ADD_USER_TO_POLL_ANSWER = gql`
-  mutation addUserToPollAnswer($pollAnswerId: String!, userId: String!) {
-    addUserToPollAnswer(pollAnswerId: $pollAnswerId, userId: $userId) { }
+  mutation addSubmissionToPollAnswer($pollAnswerId: String!, userId: String!) {
+    addSubmissionToPollAnswer(pollAnswerId: $pollAnswerId, userId: $userId) { }
   }
 `;
