+++ edtrio-server/src/server/index.ts
@@ -1,0 +1,11 @@
+import { mutations } from "./mutations";
+import { queries } from "./queries";
+import { subscriptions } from "./subscriptions";
+import { typeResolvers } from "./typeResolvers";
+
+export const resolvers = {
+  ...queries,
+  ...mutations,
+  ...subscriptions,
+  ...typeResolvers,
+};
