+++ edtrio-server/src/server/typeResolvers.ts
@@ -1,0 +1,1 @@
+export const typeResolvers = {};
